:recycle: move count
