+++ src/handlers.rs
@@ -1,5 +1,4 @@
 pub mod auth;
 pub mod cookies;
-pub mod count;
 pub mod features_i_tried;
 pub mod tasks;
